feat: add `auth` to api

+++ pages/api/visits/[id].ts
@@ -2,6 +2,7 @@ import type { NextApiRequest, NextApiResponse } from 'next'
 import _ from 'lodash'
 import { IVisit, Visit, VisitKeys } from '@database/models/visit'
 import dbConnect from '@database/dbConnect'
+import { getSession } from 'next-auth/react'
 
 dbConnect()
 
@@ -9,26 +10,38 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<IVisit>
 ) {
+  const session = await getSession({ req })
   const {
     query: { id },
     method,
+    body,
   } = req
 
   switch (method) {
     case 'GET':
-      const visit: IVisit = _.pick(
-        await Visit.findOne({ _id: id }).exec(),
-        VisitKeys
-      ) as IVisit
-      res.status(200).json(visit)
-      return
+      if (session?.user?.email == process.env.AUTHORIZED_EMAIL) {
+        const visit: IVisit = _.pick(
+          await Visit.findOne({ _id: id }).exec(),
+          VisitKeys
+        ) as IVisit
+        res.status(200).json(visit)
+        return
+      } else {
+        res.status(401).end()
+        return
+      }
 
     case 'DELETE':
-      await Visit.deleteOne({ _id: id }).exec()
-      res.status(202).end()
-      return
+      if (session?.user?.email == process.env.AUTHORIZED_EMAIL) {
+        await Visit.deleteOne({ _id: id }).exec()
+        res.status(202).end()
+        return
+      } else {
+        res.status(401).end()
+        return
+      }
   }
 
-  res.setHeader('Allow', ['GET', 'DELETE'])
+  res.setHeader('Allow', ['GET', 'DELETE', 'PUT'])
   res.status(405).end(`Method ${method} Not Allowed`)
 }
