+++ pages/api/visits/index.ts
@@ -3,6 +3,7 @@ import { ICreateVisit, IVisit, Visit, VisitKeys } from '@database/models/visit'
 import dbConnect from '@database/dbConnect'
 import _ from 'lodash'
 import { addVisit } from './addVisit'
+import { getSession } from 'next-auth/react'
 
 dbConnect()
 type Data = any
@@ -12,15 +13,22 @@ export default async function handler(
   res: NextApiResponse<Data>
 ) {
   const { method, body } = req
+  const session = await getSession({ req })
 
   switch (method) {
     case 'GET':
-      const rawVisits: IVisit[] = await Visit.find({}).exec()
-      const visits: IVisit[] = rawVisits.map((visit) =>
-        _.pick(visit, VisitKeys)
-      ) as IVisit[]
-      res.status(200).json({ plannedVisits: visits })
-      return
+      if (session?.user?.email == process.env.AUTHORIZED_EMAIL) {
+        const rawVisits: IVisit[] = await Visit.find({}).exec()
+        const visits: IVisit[] = rawVisits.map((visit) =>
+          _.pick(visit, VisitKeys)
+        ) as IVisit[]
+
+        res.status(200).json(visits)
+        return
+      } else {
+        res.status(401).end()
+        return
+      }
 
     case 'POST':
       let code = 406
@@ -41,9 +49,14 @@ export default async function handler(
       return
 
     case 'DELETE':
-      Visit.deleteMany({}).exec()
-      res.status(202).end()
-      return
+      if (session?.user?.email == process.env.AUTHORIZED_EMAIL) {
+        Visit.deleteMany({}).exec()
+        res.status(202).end()
+        return
+      } else {
+        res.status(401).end()
+        return
+      }
 
     default:
       res.setHeader('Allow', ['GET', 'POST', 'DELETE'])
